feat(precompiles): implement sealed trait pattern for PrecompileVerifier
- Removes blanket implementation of PrecompileVerifier for function types in core
- Move PrecompileVerifier trait and associated types to stdlib with sealed pattern
- Create private Sealed marker trait to restrict implementations to stdlib
- Implement KeccakVerifier struct to replace function-based verifier
- Add Sealed bound to PrecompileVerifier trait
